fix: Move config out
